+++ src/gui/nodal/SignalOperationNode.cpp
@@ -42,13 +42,13 @@ void SignalOperationNode::displayPreview()
     getOperation()->validateGraph();
     std::array<float, 100> buf;
     s_imgui_sampler_set_count(100);
-    for(size_t i=0; i<100; ++i) buf[i] = s_imgui_sampler(getOperation(), i);
+    for(int i=0; i<100; ++i) buf[i] = s_imgui_sampler(getOperation(), i);
     ImGui::PlotLines("##preview", buf.data(), 100, 0, NULL, FLT_MAX, FLT_MAX, ImVec2(0, 60.0f));
 }
 
 void SignalOperationNode::s_imgui_sampler_set_count(int count)
 {
-    qb::imgui::s_sampler_count = count;
+    qb::imgui::s_sampler_count = (float)count;
 }
 
 //--------------------------------------------------------------
@@ -59,17 +59,17 @@ float SignalOperationNode::s_imgui_sampler(void* data, int idx)
     time.duration = 1.0;
     time.t = (float)idx/qb::imgui::s_sampler_count;
     time.sec = (float)idx/qb::imgui::s_sampler_count;
-    time.elapsed = 0.01;
+    time.elapsed = 0.01f;
     return op.sample(0, time).fvec[0];
 }
 
 void SignalOperationNode::drawPreview(const Rect& previewArea)
 {
-    const float sample_count = _preview.size();
+    const int sample_count = (int)_preview.size();
     RenderInterface::setColor(0x777777FF);
     Rect target = previewArea + parentPosition + position;
     RenderInterface::fill(target.p0, target.p1);
-    vec2 barS(previewArea.size().x / sample_count, 0);
+    vec2 barS(previewArea.size().x / (float)sample_count, 0);
     RenderInterface::setColor(0xEEEEEEFF);
 
     if(_hasChange)
@@ -79,7 +79,7 @@ void SignalOperationNode::drawPreview(const Rect& previewArea)
         s_imgui_sampler_set_count(sample_count);
         for(size_t i=0; i<sample_count; ++i)
         {
-            _preview[i] = s_imgui_sampler(getOperation(), i);
+            _preview[i] = s_imgui_sampler(getOperation(), (int)i);
             if (_preview[i] < minVal) minVal = _preview[i];
             if (_preview[i] > maxVal) maxVal = _preview[i];
         }
@@ -89,9 +89,9 @@ void SignalOperationNode::drawPreview(const Rect& previewArea)
 
     for(int i=0; i< sample_count; ++i)
     {
-        vec2 p = target.p0 + vec2(i*barS.x,previewArea.size().y*0.5);
+        vec2 p = target.p0 + vec2((float)i*barS.x,previewArea.size().y*0.5f);
         float y = _preview[i] / _previewMaxVal;
-        float h = previewArea.size().y*0.5 * -y;
+        float h = previewArea.size().y*0.5f * -y;
         RenderInterface::fill(p, p+vec2(barS.x,h));
     }
 }
